% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_get_domain_association}
\alias{amplify_get_domain_association}
\title{Retrieves domain info that corresponds to an appId and domainName}
\usage{
amplify_get_domain_association(appId, domainName)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{domainName}{[required] Name of the domain.}
}
\description{
Retrieves domain info that corresponds to an appId and domainName.
}
\section{Request syntax}{
\preformatted{svc$get_domain_association(
  appId = "string",
  domainName = "string"
)
}
}

\keyword{internal}
