% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_update_app}
\alias{amplify_update_app}
\title{Updates an existing Amplify App}
\usage{
amplify_update_app(appId, name, description, platform,
  iamServiceRoleArn, environmentVariables, enableBranchAutoBuild,
  enableBasicAuth, basicAuthCredentials, customRules, buildSpec)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{name}{Name for an Amplify App.}

\item{description}{Description for an Amplify App.}

\item{platform}{Platform for an Amplify App.}

\item{iamServiceRoleArn}{IAM service role for an Amplify App.}

\item{environmentVariables}{Environment Variables for an Amplify App.}

\item{enableBranchAutoBuild}{Enables branch auto-building for an Amplify App.}

\item{enableBasicAuth}{Enables Basic Authorization for an Amplify App.}

\item{basicAuthCredentials}{Basic Authorization credentials for an Amplify App.}

\item{customRules}{Custom redirect / rewrite rules for an Amplify App.}

\item{buildSpec}{BuildSpec for an Amplify App.}
}
\description{
Updates an existing Amplify App.
}
\section{Request syntax}{
\preformatted{amplify$update_app(
  appId = "string",
  name = "string",
  description = "string",
  platform = "IOS"|"ANDROID"|"WEB"|"REACT_NATIVE",
  iamServiceRoleArn = "string",
  environmentVariables = list(
    "string"
  ),
  enableBranchAutoBuild = TRUE|FALSE,
  enableBasicAuth = TRUE|FALSE,
  basicAuthCredentials = "string",
  customRules = list(
    list(
      source = "string",
      target = "string",
      status = "string",
      condition = "string"
    )
  ),
  buildSpec = "string"
)
}
}

\keyword{internal}
