% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_delete_function}
\alias{appsync_delete_function}
\title{Deletes a Function}
\usage{
appsync_delete_function(apiId, functionId)
}
\arguments{
\item{apiId}{[required] The GraphQL API ID.}

\item{functionId}{[required] The \code{Function} ID.}
}
\description{
Deletes a \code{Function}.
}
\section{Request syntax}{
\preformatted{appsync$delete_function(
  apiId = "string",
  functionId = "string"
)
}
}

\keyword{internal}
