% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_get_resolver}
\alias{appsync_get_resolver}
\title{Retrieves a Resolver object}
\usage{
appsync_get_resolver(apiId, typeName, fieldName)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{typeName}{[required] The resolver type name.}

\item{fieldName}{[required] The resolver field name.}
}
\description{
Retrieves a \code{Resolver} object.
}
\section{Request syntax}{
\preformatted{appsync$get_resolver(
  apiId = "string",
  typeName = "string",
  fieldName = "string"
)
}
}

\keyword{internal}
