% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_list_resolvers_by_function}
\alias{appsync_list_resolvers_by_function}
\title{List the resolvers that are associated with a specific function}
\usage{
appsync_list_resolvers_by_function(apiId, functionId, nextToken,
  maxResults)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{functionId}{[required] The Function ID.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which you can use to return the next set of items in the
list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\description{
List the resolvers that are associated with a specific function.
}
\section{Request syntax}{
\preformatted{svc$list_resolvers_by_function(
  apiId = "string",
  functionId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
