% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_device_pool}
\alias{devicefarm_get_device_pool}
\title{Gets information about a device pool}
\usage{
devicefarm_get_device_pool(arn)
}
\arguments{
\item{arn}{[required] The device pool's ARN.}
}
\description{
Gets information about a device pool.
}
\section{Request syntax}{
\preformatted{svc$get_device_pool(
  arn = "string"
)
}
}

\examples{
# The following example returns information about a specific device pool,
# given a project ARN.
\donttest{svc$get_device_pool(
  arn = "arn:aws:devicefarm:us-west-2:123456789101:project:EXAMPLE-GUID-123-456"
)}

}
\keyword{internal}
