% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_get_network_profile}
\alias{devicefarm_get_network_profile}
\title{Returns information about a network profile}
\usage{
devicefarm_get_network_profile(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the network profile you want to return
information about.}
}
\description{
Returns information about a network profile.
}
\section{Request syntax}{
\preformatted{svc$get_network_profile(
  arn = "string"
)
}
}

\keyword{internal}
