% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devicefarm_operations.R
\name{devicefarm_update_network_profile}
\alias{devicefarm_update_network_profile}
\title{Updates the network profile with specific settings}
\usage{
devicefarm_update_network_profile(arn, name, description, type,
  uplinkBandwidthBits, downlinkBandwidthBits, uplinkDelayMs,
  downlinkDelayMs, uplinkJitterMs, downlinkJitterMs, uplinkLossPercent,
  downlinkLossPercent)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the project for which you want to
update network profile settings.}

\item{name}{The name of the network profile about which you are returning
information.}

\item{description}{The descriptoin of the network profile about which you are returning
information.}

\item{type}{The type of network profile you wish to return information about. Valid
values are listed below.}

\item{uplinkBandwidthBits}{The data throughput rate in bits per second, as an integer from 0 to
104857600.}

\item{downlinkBandwidthBits}{The data throughput rate in bits per second, as an integer from 0 to
104857600.}

\item{uplinkDelayMs}{Delay time for all packets to destination in milliseconds as an integer
from 0 to 2000.}

\item{downlinkDelayMs}{Delay time for all packets to destination in milliseconds as an integer
from 0 to 2000.}

\item{uplinkJitterMs}{Time variation in the delay of received packets in milliseconds as an
integer from 0 to 2000.}

\item{downlinkJitterMs}{Time variation in the delay of received packets in milliseconds as an
integer from 0 to 2000.}

\item{uplinkLossPercent}{Proportion of transmitted packets that fail to arrive from 0 to 100
percent.}

\item{downlinkLossPercent}{Proportion of received packets that fail to arrive from 0 to 100
percent.}
}
\description{
Updates the network profile with specific settings.
}
\section{Request syntax}{
\preformatted{devicefarm$update_network_profile(
  arn = "string",
  name = "string",
  description = "string",
  type = "CURATED"|"PRIVATE",
  uplinkBandwidthBits = 123,
  downlinkBandwidthBits = 123,
  uplinkDelayMs = 123,
  downlinkDelayMs = 123,
  uplinkJitterMs = 123,
  downlinkJitterMs = 123,
  uplinkLossPercent = 123,
  downlinkLossPercent = 123
)
}
}

\keyword{internal}
