% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_operations.R
\name{mobile_export_project}
\alias{mobile_export_project}
\title{Exports project configuration to a snapshot which can be downloaded and
shared}
\usage{
mobile_export_project(projectId)
}
\arguments{
\item{projectId}{[required] Unique project identifier.}
}
\description{
Exports project configuration to a snapshot which can be downloaded and
shared. Note that mobile app push credentials are encrypted in exported
projects, so they can only be shared successfully within the same AWS
account.
}
\section{Request syntax}{
\preformatted{mobile$export_project(
  projectId = "string"
)
}
}

\keyword{internal}
