% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileanalytics_service.R
\name{mobileanalytics}
\alias{mobileanalytics}
\title{Amazon Mobile Analytics}
\usage{
mobileanalytics()
}
\description{
Amazon Mobile Analytics is a service for collecting, visualizing, and
understanding app usage data at scale.
}
\section{Example}{
\preformatted{mobileanalytics <- paws::mobileanalytics()
mobileanalytics$operation()
}
}

\section{Operations}{

\tabular{ll}{
\link[=mobileanalytics_put_events]{put_events} \tab The PutEvents operation records one or more events
}
}

